Hooks up enum value parsing
